Move fake Gnocchi client to "gnocchi/testhelper/client/fake.go"
That scheme would match "gophercloud/gophercloud".
